+++ GameFrameworks/GameFrameworks/ConfigUtil.h
@@ -13,9 +13,9 @@ namespace meltshine
 		static void Write(const TCHAR* section, const TCHAR* key, const TCHAR* str);
 		static void Write(const TCHAR* section, const TCHAR* key, const int& num);
 #ifdef _UNICODE
-		static std::wstring ReadString(const TCHAR* section, const TCHAR* key, const size_t& size);
+		static std::wstring ReadString(const TCHAR* section, const TCHAR* key, const DWORD& size);
 #else
-		static std::string ReadString(const TCHAR* section, const TCHAR* key, const size_t& size);
+		static std::string ReadString(const TCHAR* section, const TCHAR* key, const DWORD& size);
 #endif // _UNICODE
 
 		static int ReadInt(const TCHAR* section, const TCHAR* key);
